pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "governance-token-analyzer"
version = "0.1.0"
authors = [
    {name = "The Authors", email = "author@example.com"},
]
description = "A tool for analyzing governance token distributions across DeFi protocols"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "requests>=2.31.0",
    "pandas>=2.0.3",
    "numpy>=1.24.4",
    "matplotlib>=3.7.2",
    "seaborn>=0.12.2",
    "python-dotenv>=1.0.0",
    "click>=8.1.3",
    "jinja2>=3.1.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "ruff>=0.1.5",
    "mypy>=1.6.1",
    "pytest-cov>=4.1.0",
]

[project.scripts]
gta = "governance_token_analyzer.cli:cli"

[project.urls]
"Homepage" = "https://github.com/your-username/governance-token-distribution-analyzer"
"Bug Tracker" = "https://github.com/your-username/governance-token-distribution-analyzer/issues"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 120
target-version = "py38"

[tool.ruff.lint]
# Keep critical errors that affect functionality
select = [
    "E",       # pycodestyle errors
    "F",       # pyflakes errors  
    "B",       # flake8-bugbear
    "C4",      # flake8-comprehensions
    "SIM",     # flake8-simplify
    "PTH",     # flake8-use-pathlib
    "D",       # pydocstyle
]

# Ignore non-critical style issues to focus on functionality
ignore = [
    "E501",    # line-too-long (cosmetic)
    "PTH118",  # os-path-join (modernization, not breaking)
    "PTH123",  # builtin-open (modernization, not breaking) 
    "PTH103",  # os-makedirs (modernization, not breaking)
    "PTH120",  # os-path-dirname (modernization, not breaking)
    "PTH110",  # os-path-exists (modernization, not breaking)
    "PTH100",  # os-path-abspath (modernization, not breaking)
    "PTH101",  # os-chmod (modernization, not breaking)
    "PTH106",  # os-rmdir (modernization, not breaking)
    "PTH107",  # os-remove (modernization, not breaking)
    "PTH111",  # os-path-expanduser (modernization, not breaking)
    "PTH119",  # os-path-basename (modernization, not breaking)
    "PTH202",  # os-path-getsize (modernization, not breaking)
    "D205",    # blank-line-after-summary (docs formatting)
    "D100",    # undocumented-public-module (docs)
    "D103",    # undocumented-public-function (docs)
    "D107",    # undocumented-public-init (docs)
    "D200",    # fits-on-one-line (docs formatting)
    "D415",    # ends-in-punctuation (docs formatting)
    "B007",    # unused-loop-control-variable (style preference)
    "SIM117",  # multiple-with-statements (style preference)
    "SIM105",  # suppressible-exception (style preference)
]

per-file-ignores = { "tests/*" = ["D100", "D103", "E402"] }

[tool.ruff.format]
# Enable the formatter
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false 